requirements.txt
asn1crypto==1.4.0
astroid==2.11.5
asttokens==2.0.5
async-generator==1.10
attrs
autograd==1.3
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.2
black==22.1.0
boto3
botocore==1.23.13
cachetools==4.2.4
cdifflib==1.2.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.4
colorama==0.4.3
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
dill==0.3.4
docutils==0.15.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
fire==0.4.0
fonttools==4.28.2
future==0.18.2
fuzzywuzzy==0.18.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.2.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
gspread==4.0.1
gspread-dataframe==3.3.0
gspread-pandas==2.3.0
h11==0.13.0
httplib2==0.20.2
idna==3.3
watchtower==1.0.6
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
Lifetimes==0.11.3
MarkupSafe
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.7.0
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.0
nest-asyncio==1.5.4
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.9
oscrypto==1.2.1
outcome==1.1.0
packaging==21.3
pandas==1.3.4
paramiko==2.11.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
protobuf==3.19.1
psycopg2==2.9.2
pure-eval==0.2.2
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.14.1
Pygments==2.11.2

pylint==2.14.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==3.0.6
PySocks==1.7.1
pyspark==3.2.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-Levenshtein==0.12.2
pytz==2021.3

pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0

SQLAlchemy
sqlparse
stack-data

